Changed rest of the imports and moved stopwatch utility function to general functions
