+++ modules/general_functions.py
@@ -40,15 +40,40 @@ import shutil
 import subprocess
 import tempfile
 import logging
+import time
 
+from timeit import default_timer as timer
 from pathlib import Path
 from decimal import Decimal
-from modules.parsing import ToFListParser
+from .parsing import ToFListParser
 from subprocess import Popen
-from tests.utils import stopwatch
 
 
 # TODO this could still be organized into smaller modules
+
+def stopwatch(log_file=None):
+    """Decorator that measures the time it takes to execute a function
+    and prints the results or writes them to a log file if one is provided
+    as an argument.
+    """
+    def outer(func):
+        def inner(*args, **kwargs):
+            start = timer()
+            res = func(*args, **kwargs)
+            stop = timer()
+
+            timestamp = time.strftime("%y/%m/%D %H:%M.%S")
+            msg = f"{timestamp}: {func.__name__}({args, kwargs})\n\t" \
+                  f"took {stop - start} to execute.\n"
+            if log_file is None:
+                print(msg)
+            else:
+                with open(log_file, "a") as file:
+                    file.write(msg)
+            return res
+        return inner
+    return outer
+
 
 def rename_file(old_path, new_name):
     """Renames file or directory and returns new path.
